fix(sqlgen,cdc,federated): EAV numeric storage width is DOUBLE; round-trip float64 operands; one bool operand rule (#384)
Resolves the four findings of the second PR #500 review:

P1a — the BIGINT projection for EAV-only integer/smallint mangled
non-integral history (TRY_CAST(1.5 AS BIGINT) = 2), matching equals:2 only
on the DuckDB legs while Postgres compared the raw NUMERIC 1.5. The true
storage width is DOUBLE: the write funnel narrows every numeric-family
value through float64 (numutil.Float64), so value_numeric only ever holds
float64 images and DOUBLE reproduces all of them — over-range and
non-integral alike — exactly. All four projection surfaces (hot pivot,
list elements, CDC export, cold NULL augmentation) move from BIGINT to
DOUBLE in lockstep, joining the numeric class.

P2a — integer/smallint operands follow to CAST(? AS DOUBLE), so an
operand of any magnitude (1e30) compares on both engines instead of
raising a DuckDB Conversion Error. The bigint sibling (BIGINT must stay
exact above 2^53) is pre-existing and tracked as #502.

P1b — decimalString now renders float64 operands in shortest round-trip
form (strconv.FormatFloat 'g' -1); %.15g dropped the 16th-17th
significant digits, so DuckDB bound an adjacent float64 and equality
missed rows only on that route.

P2b — bool operands parse under one engine-shared rule (parseBoolOperand:
ParseBool spellings, else any integer at the PG path's historical >0
truthiness). Before, "2" bound only on PG and "true" only on DuckDB.

The acceptance e2e now drives every probe through hot, warm (real CDC
flush), and cold (real compaction) and adds non-integral-history,
16-digit numeric, bool-spelling, and past-BIGINT-operand probes. Suite
shape pins and docs §6.4 updated to the DOUBLE contract.

+++ internal/e2e_harness/production/full_type_parquet_e2e_test.go
@@ -18,8 +18,9 @@ import (
 // castEAVValue (internal/cdc/duckdb_exporter.go); note the deliberate
 // asymmetries: bound uuid (ref) is physical UUID while EAV uuid (token) is
 // VARCHAR, and bound smallint/integer (rank/count) keep declared width while
-// EAV-only level/qty export at storage width BIGINT (#384 — value_numeric is
-// unconstrained NUMERIC, so declared width manufactured NULLs).
+// EAV-only level/qty export at storage width DOUBLE (#384 — value_numeric
+// holds float64 images from the write funnel, so declared width manufactured
+// NULLs for over-range history and rounded non-integral history).
 var wideParquetTypes = map[string]string{
 	"schema_id": "SMALLINT", "row_id": "UUID",
 	"changed_at": "BIGINT", "deleted_at": "BIGINT",
@@ -27,7 +28,7 @@ var wideParquetTypes = map[string]string{
 	"title": "VARCHAR", "rank": "SMALLINT", "count": "INTEGER", "amount": "BIGINT",
 	"score": "DOUBLE", "ref": "UUID", "joined": "BIGINT", "touched": "BIGINT",
 	"note": "VARCHAR", "active": "BOOLEAN", "born": "BIGINT", "seen": "BIGINT",
-	"level": "BIGINT", "qty": "BIGINT", "total": "BIGINT", "ratio": "DOUBLE", "token": "VARCHAR",
+	"level": "DOUBLE", "qty": "DOUBLE", "total": "BIGINT", "ratio": "DOUBLE", "token": "VARCHAR",
 	// #204: list attrs export as a DuckDB LIST of the items type.
 	"tags": "VARCHAR[]",
 }
